+++ LostInTheMap/map_system.cpp
@@ -135,8 +135,8 @@ int ** map_system::get_pathfinding_map(ITerrain * tilemap)
 SDL_Point map_system::world_to_tilemap_ids(SDL_Point world_coords, ITerrain* tilemap)
 {
 	SDL_Point result;
-	result.x = std::floor(world_coords.x / tilemap->tile_width);
-	result.y = std::floor(world_coords.y / tilemap->tile_width);
+	result.x = (int)std::floor(world_coords.x / tilemap->tile_width);
+	result.y = (int)std::floor(world_coords.y / tilemap->tile_width);
 	return result;
 }
